Extracted dedicated UserDetails component
